+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.1.3
+
+* Fix - mam wrapping improved for fetch
+* Fix - mam.subscribe updated to match @iota/mam signature
+
 ## v0.1.2
 
 * Fix - wrapMethodCallbackOrAsync updated to use methodsName for getTrytes check